s/Conviction/Disposition/
Update model names and calling code to reflect the realization that
records in our data are dispositions, not convictions.

This initially seemed like extra work, but we'll need an actual
Conviction model which we'll extract from the Disposition models.
This will make the queries a lot easier and faster.

Addresses https://github.com/sc3/cook-convictions/issues/41