Add function to remove attempted offense statute
Add a function to remove the statutes for attempted offenses
from statute citation strings.  This should make it easier to
parse statutes for looking up IUCR codes.

Add a management command to export a CSV of all statutes that
seem to have the attempted offense statutes in them to help
prepare test cases.